apps/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
sqlalchemy = "*"
asyncpg = "*"
alembic = "*"
pydantic-settings = "^2.13.1"
psycopg2-binary = "*"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I"]
